Partial fix memory issues
Use strcpy for string copying
Added type casting for clarity and NULL setting for safety
Fix scope memory allocation typo
Use new error code range

+++ src/scope.h
@@ -32,7 +32,7 @@
 typedef struct Scope {
 	struct Scope* parentScope;
 	hashtable_t* variables;
-	int storageSize;
+	size_t storageSize;
 	char** storedIdentifiers;
 	int identifierCount;
 } Scope;
